no anda el test pero si el llamado

+++ Kernel/include/memory_management.h
@@ -1,9 +1,10 @@
 #ifndef MEMORY_MANAGEMENT_H
 #define MEMORY_MANAGEMENT_H
 
+#include <stdint.h>
 
-void my_mm_init(void *p, int s);
-void *my_malloc(int size);
+void my_mm_init(void *p, uint64_t s);
+void *my_malloc(uint64_t size);
 void my_free(void *p);
