+++ Kernel/libs/memory_management.c
@@ -28,7 +28,7 @@ int size;
 int current;
 void *free_ptrs[BLOCK_COUNT]; // [10, 20, 30, 40, 50] --> [20, 30, 40, 50] --> [30, 40, 50]
 
-void my_mm_init(void *p, int s)
+void my_mm_init(void *p, uint64_t s)
 {
     start = p;
     size = s;
@@ -41,7 +41,7 @@ void my_mm_init(void *p, int s)
     current = 0;
 }
 
-void *my_malloc(int size)
+void *my_malloc(uint64_t size)
 {
     if (size > BLOCK_SIZE || current >= BLOCK_COUNT)
     {
